+++ core/utils/insect_attributes.py
