+++ core/wagtail_hooks.py
@@ -3,6 +3,9 @@ from django.templatetags.static import static
 
 from wagtail import hooks
 
-@hooks.register('insert_global_admin_css')
+
+@hooks.register("insert_global_admin_css")
 def global_admin_css():
-    return format_html('<link rel="stylesheet" href="{}">', static('css/memcollection.css'))
+    return format_html(
+        '<link rel="stylesheet" href="{}">', static("css/memcollection.css")
+    )
