+++ images/views.py
@@ -1,5 +1,6 @@
 from wagtail.api.v2.views import BaseAPIViewSet
 
+from core.utils.helpers import get_fields
 from images.models import HabitatImage, InsectImage, PlantImage, SpecimenRecordImage
 from images.serializers import (
     HabitatImageSerializer,
@@ -7,7 +8,6 @@ from images.serializers import (
     PlantImageSerializer,
     SpecimenRecordImageSerializer,
 )
-from utils.helpers import get_fields
 
 
 class SpecimenRecordImagesAPIViewSet(BaseAPIViewSet):
